requirements.txt
selenium>=3.141.0
PyYAML>=3.13
InstaPy

# api-display-purposes>=0.0.3; python_version > '3.5'
# certifi>=2018.10.15
# chardet>=3.0.4,<4
# clarifai>=2.4.1
# configparser>=3.5.0,<4
# EasyProcess>=0.2.3
# emoji>=1.6.0
# future>=0.17.1
# grpcio>=1.16.1
# idna>=2.7,<3
# jsonschema>=2.6.0,<3
# MeaningCloud-python>=1.1.1
# plyer>=1.3.1
# protobuf>=3.6.1,<4
# python-telegram-bot>=12.0.0; python_version > '3.6'
# PyVirtualDisplay>=0.2.1; sys_platform != 'win32'
# PyYAML>=3.13
# regex>=2018.11.22
# requests>=2.20.1
# selenium>=3.141.0
# setuptools_rust>=0.11.6
# six>=1.11.0
# urllib3>=1.24.1
# webdriverdownloader>=1.1.0.3